Fixes #2
